moved opening chart to charts. renamed to charts
