+++ src/components/Charts/OpeningChart.vue
